LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/usemap-casing-expected.txt


PASS Image with usemap of #sanityCheck should match the area with map named sanityCheck 
FAIL Image with usemap of #sImPlE should not match any of the areas assert_equals: The element retrieved must be the image, not an area; used <map> with name "simple" expected Element node <img src="/images/threecolors.png" usemap="#sImPlE" width... but got Element node <area shape="rect" coords="0,0,300,300"></area>
FAIL Image with usemap of #paSSfield-killroyß should not match any of the areas assert_equals: The element retrieved must be the image, not an area; used <map> with name "passfield-killroyß" expected Element node <img src="/images/threecolors.png" usemap="#paSSfield-kil... but got Element node <area shape="rect" coords="0,0,300,300"></area>
FAIL Image with usemap of #глупый should not match any of the areas assert_equals: The element retrieved must be the image, not an area; used <map> with name "ГЛУПЫЙ" expected Element node <img src="/images/threecolors.png" usemap="#глупый" width... but got Element node <area shape="rect" coords="0,0,300,300"></area>
FAIL Image with usemap of #åωk should not match any of the areas null is not an object (evaluating 'element.parentElement')
FAIL Image with usemap of #blah1 should not match any of the areas null is not an object (evaluating 'element.parentElement')
FAIL Image with usemap of #tÉdz5アパートFi should not match any of the areas null is not an object (evaluating 'element.parentElement')
FAIL Image with usemap of #ΣΣ should not match any of the areas null is not an object (evaluating 'element.parentElement')
           
